+++ routes.php
@@ -1,4 +1,3 @@
 <?php
 
 Route::post('/sir-trevor/upload', 'ImageController@upload');
-Route::get('/sir-trevor/upload', 'ImageController@upload');