+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+## Updated
+- Moved ImageController from Http to src
+
 # [1.0.0]
 ## Updated
 - composer.json for 1.0.0